requirements.txt
streamlit==1.24.0
pandas==2.0.3
numpy==1.24.3
openai==0.27.8
python-dotenv==1.0.0
PyPDF2==3.0.1
sentence-transformers==2.2.2
scikit-learn==1.3.0